:truck: [#175] -- Split views per business domain
