+++ setup.cfg
@@ -1,7 +1,7 @@
 [pycodestyle]
 [pep8]
 [flake8]
-ignore=W293,W291,E501,E261
+ignore=E501,E261,W291,W293,W503
 max-line-length=88
 exclude=migrations,static,media
