+++ src/brouwers/shop/views/__init__.py
@@ -1,0 +1,21 @@
+from .dev import OrderConfirmationEmailView
+from .public import (
+    CartDetailView,
+    CategoryDetailView,
+    CheckoutView,
+    ConfirmOrderView,
+    IndexView,
+    ProductDetailView,
+    RouterView,
+)
+
+__all__ = [
+    "IndexView",
+    "RouterView",
+    "CategoryDetailView",
+    "ProductDetailView",
+    "CartDetailView",
+    "CheckoutView",
+    "ConfirmOrderView",
+    "OrderConfirmationEmailView",
+]
